Consolidate binding serialization scripts a little bit
