+++ Assets/Script/Input/Bindings/Serialization/BindingSerialization.cs
@@ -4,19 +4,54 @@ using System.Security.Cryptography;
 using System.Text;
 using System.Text.RegularExpressions;
 using UnityEngine.InputSystem;
+using YARG.Audio;
+using YARG.Core;
 
 namespace YARG.Input.Serialization
 {
-    public static class DeviceSerializationExtensions
+    using BindingCollection = Dictionary<string, List<SerializedInputControl>>;
+
+    public class SerializedProfileBindings
     {
-        private static readonly Regex _xinputUserIndexRegex = new(@"\\""userIndex\\"":\s*\d,");
+        public List<SerializedInputDevice> Devices = new();
+        public SerializedMic Microphone;
+
+        public Dictionary<GameMode, BindingCollection> Bindings = new();
+        public BindingCollection MenuBindings = new();
+    }
+
+    public class SerializedInputDevice
+    {
+        public string Layout;
+        public string Hash;
+
+        public bool MatchesDevice(InputDevice device)
+        {
+            return Layout == device.layout && Hash == device.GetHash();
+        }
+    }
+
+    public class SerializedInputControl
+    {
+        public SerializedInputDevice Device;
+        public string ControlPath;
+        public Dictionary<string, string> Parameters = new();
+    }
+
+    public static class BindingSerialization
+    {
         private static readonly SHA1 _hashAlgorithm = SHA1.Create();
+        private static readonly Regex _xinputUserIndexRegex = new(@"\\""userIndex\\"":\s*\d,");
 
         private static readonly Dictionary<InputDevice, string> _hashCache = new();
 
         public static SerializedInputDevice Serialize(this InputDevice device)
         {
-            return SerializedInputDevice.Serialize(device);
+            return new()
+            {
+                Layout = device.layout,
+                Hash = device.GetHash(),
+            };
         }
 
         public static string GetHash(this InputDevice device)
