--- Assets/Script/Input/Bindings/Serialization/SerializedBindings.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f1ac81c5568489b4d90754c9eb1353e8
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
